Create report-http-status-codes.php
If you manage a windows server and have many redirects this script will parse the web.config and email you a list of redirect URLS sorted by corresponding status code (except 200). I run it as a scheduled task so I can maintain the large number of redirects on the server. It could certainly use some tweaks so don't be afraid to hack it up.